% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smiles_similarity_search.R
\name{smiles_similarity_search}
\alias{smiles_similarity_search}
\title{Similarity search using smiles}
\usage{
smiles_similarity_search(sim_db, counts_db, smiles, threshold = 0.7)
}
\arguments{
\item{sim_db}{Similarity database}

\item{counts_db}{fingerprint counts database}

\item{smiles}{smiles to search}

\item{threshold}{tanimoto similarity threshold}
}
\value{
data frame with sim_db id, chemical ID, smiles of hit, similarity score
}
\description{
Similarity search using smiles
}
\examples{

\dontrun{

smiles <- "CC(=O)OC1=CC=CC=C1C(=O)O"

sim_db <- mongolite::mongo(db="chemdb",url = "mongodb://localhost:27017" ,collection = "molecules")

counts_db <- mongolite::mongo(db="chemdb",url = "mongodb://localhost:27017" ,collection = "mfp_counts")
smiles_similarity_search(sim_db,counts_db,smiles, threshold = 0.7)
}

}
