% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smiles_substructure_search.R
\name{smiles_substructure_search}
\alias{smiles_substructure_search}
\title{Substructure search using smiles}
\usage{
smiles_substructure_search(
  sim_db,
  smiles,
  max_mismatches = 2,
  threshold = 0.9,
  min_overlap_coefficient = 0.98,
  au = 2,
  bu = 1,
  numParallel = 5
)
}
\arguments{
\item{sim_db}{Similarity database}

\item{smiles}{smiles to search}

\item{max_mismatches}{max number of mismatched fingerprint bits}

\item{threshold}{%overlap required for inital search}

\item{numParallel}{number of processors parameter for fmcsBatch final similarity}

\item{counts_db}{fingerprint counts database}

\item{au=2}{parameter for fmcsBatch final similarity}

\item{bu=1}{parameter for fmcsBatch final similarity}
}
\value{
data frame with Query_Size,Target_Size,MCS_Size,Tanimoto_Coefficient,Overlap_Coefficient,id
}
\description{
Substructure search using smiles
}
\examples{

\dontrun{

smiles <- "CC(=O)OC1=CC=CC=C1C(=O)O"

sim_db <- mongolite::mongo(db="chemdb",url = "mongodb://localhost:27017" ,collection = "similarity")


smiles_substructure_search(sim_db_db,smiles)
}

}
